Temporary fix for #21
We cannot create the window inside device creation as the loader
cannot handle any wsi calls in there.
This means we cannot query queue support for the window.
We just guess for now, but could do better in future.

NOTE: this also breaks the standalone device mode. To be fixed!